Add tests for internal linking
