+++ lib/url.js
@@ -23,6 +23,34 @@ export function removeTracking (value) {
   return value
 }
 
+/**
+ * parse links like https://stacker.news/items/123456 as #123456
+ */
+export function parseInternalLinks (href) {
+  const url = new URL(href)
+  const { pathname, searchParams } = url
+  // ignore empty parts which exist due to pathname starting with '/'
+  const emptyPart = part => !!part
+  const parts = pathname.split('/').filter(emptyPart)
+  if (parts[0] === 'items' && /^[0-9]+$/.test(parts[1])) {
+    const itemId = parts[1]
+    // check for valid item page due to referral links like /items/123456/r/ekzyis
+    const itemPages = ['edit', 'ots', 'related']
+    const itemPage = itemPages.includes(parts[2]) ? parts[2] : null
+    if (itemPage) {
+      // parse   https://stacker.news/items/1/related?commentId=2
+      // as      #1/related
+      // and not #2
+      // since commentId will be ignored anyway
+      const linkText = `#${itemId}/${itemPage}`
+      return linkText
+    }
+    const commentId = searchParams.get('commentId')
+    const linkText = `#${commentId || itemId}`
+    return linkText
+  }
+}
+
 export function stripTrailingSlash (uri) {
   return uri.endsWith('/') ? uri.slice(0, -1) : uri
 }
